Get some little changes work done

+++ app/Http/Controllers/RoleController.php
@@ -3,9 +3,12 @@
 namespace App\Http\Controllers;
 
 // use App\Models\Role;
+
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Auth\Events\Validated;
 use Spatie\Permission\Models\Permission;
 
@@ -14,10 +17,10 @@ class RoleController extends Controller
 
 
     public function __construct()
-    {      
+    {
         $this->middleware('permission:Delete Permission', ['only' => ['destroy']]);
-        $this->middleware('permission:Edit Permission', ['only' => ['edit','update']]);
-        $this->middleware('permission:Create Permission', ['only' => ['create','store']]);
+        $this->middleware('permission:Edit Permission', ['only' => ['edit', 'update']]);
+        $this->middleware('permission:Create Permission', ['only' => ['create', 'store']]);
         $this->middleware('permission:View Permission', ['only' => ['view']]);
         $this->middleware('permission:Index Permission', ['only' => ['index']]);
     }
@@ -26,13 +29,16 @@ class RoleController extends Controller
      * Display a listing of the resource.
      */
     public function index()
-    {   
+    {
 
         $roles = Role::paginate(3);
-        
-        return view('roles-and-permission.roles.index'
-                ,['roles' =>  $roles,
-                ]); 
+
+        return view(
+            'roles-and-permission.roles.index',
+            [
+                'roles' =>  $roles,
+            ]
+        );
     }
 
     /**
@@ -40,8 +46,8 @@ class RoleController extends Controller
      */
     public function create()
     {
-        $permissions = DB::table('permissions')->pluck('name','name');
-        return view('roles-and-permission.roles.create',['permissions' => $permissions]); 
+        $permissions = DB::table('permissions')->pluck('name', 'name');
+        return view('roles-and-permission.roles.create', ['permissions' => $permissions]);
     }
 
     /**
@@ -49,8 +55,8 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
-        
-        
+
+
         $data = $request->validate([
             'permissions' => 'required|array',
             'name' => 'required|
@@ -60,15 +66,16 @@ class RoleController extends Controller
 
 
         $role = Role::create(
-            [ 'name' => $request->name]);
-        
+            ['name' => $request->name]
+        );
+
 
         $permissions = $request->permissions;
         $role->syncPermissions($permissions);
 
         $role->save();
 
-        return redirect('roles')->with('status','Role successfully inserted');
+        return redirect('roles')->with('status_success', 'Role successfully inserted');
     }
     /**
      * Display the specified resource.
@@ -78,6 +85,18 @@ class RoleController extends Controller
         //
     }
 
+    public function compareRole($rolename)
+    {
+        $user = Auth::user();
+        $loggedInUserRoleName = $user->roles->pluck('name', 'name');
+        foreach ($loggedInUserRoleName as $role) {
+            $loggedInUserRole = $role;
+        }
+        if ($loggedInUserRole === $rolename) {
+            return true;
+        }
+        return false;
+    }
     /**
      * Show the form for editing the specified resource.
      */
@@ -85,17 +104,24 @@ class RoleController extends Controller
     {
         $permissions = Permission::all();
         $role = Role::findOrFail($id);
-        $rolePermissions = DB::table('role_has_permissions')
-                            ->where('role_has_permissions.role_id',$role->id)
-                            ->pluck('role_has_permissions.permission_id','role_has_permissions.permission_id')
-                            ->all();
+        $compareRole = $this->compareRole($role->name);
+        if ($compareRole == true) {
+            return redirect()->route('roles.index')->with('status_error', 'You can\'t update your own role and permission');
+        } else {
+            $rolePermissions = DB::table('role_has_permissions')
+                ->where('role_has_permissions.role_id', $role->id)
+                ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
+                ->all();
 
-        return view('roles-and-permission.roles.edit', 
-                                                    ['role' => $role,
-                                                    'rolePermissions' => $rolePermissions,
-                                                    'permissions' => $permissions 
-                                                    ]
-                    );
+            return view(
+                'roles-and-permission.roles.edit',
+                [
+                    'role' => $role,
+                    'rolePermissions' => $rolePermissions,
+                    'permissions' => $permissions
+                ]
+            );
+        }
     }
 
     /**
@@ -107,18 +133,18 @@ class RoleController extends Controller
             'name' => 'required',
             'permissions' => 'required|array',
         ]);
-       
+
         $role = Role::findOrfail($id);
 
         $permissions = $request->permissions;
-       
-        
-        if($role){
+
+
+        if ($role) {
             $role->syncPermissions($permissions);
             $role->update(['name' => $request->name]);
-            return redirect('roles')->with('status','Role successfully updated');
-        }else{
-        return redirect('roles')->with('status','Something is wrong Role cannot updated please try again');
+            return redirect('roles')->with('status_success', 'Role successfully updated');
+        } else {
+            return redirect('roles')->with('status', 'Something is wrong Role cannot updated please try again');
         }
     }
 
@@ -129,19 +155,22 @@ class RoleController extends Controller
     {
         $role = Role::findOrfail($id);
 
-
-        if($role){
-            $role->delete();
-        }else{
-        return redirect('roles')->with('status','Something is wrong Role cannot be delete please try again');
+        $compareRole = $this->compareRole($role->name);
+        if ($compareRole == true) {
+            return redirect()->route('roles.index')->with('status', 'You can\'t delete your own role and permission');
+        } else {
+            if ($role) {
+                $role->delete();
+                return redirect('roles')->with('status', 'Role deleted successfully');
+            } else {
+                return redirect('roles')->with('status', 'Something is wrong Role cannot be delete please try again');
+            }
         }
-
-        return redirect('roles')->with('status','Role deleted successfully');
     }
 
 
-    public function checkRole(Request $request){
-
+    public function checkRole(Request $request)
+    {
         $roleName =  $request->input('roleName');
         $roleExists = Role::where('name', $roleName)->exists();
         return response()->json(['exists' => $roleExists]);
